refactored movie component and called API for a single movie. still needs functionality
