+++ client/src/pages/Homepage.jsx
@@ -1,10 +1,60 @@
+import { useEffect, useState } from 'react';
+import { Link } from 'react-router-dom';
 import Movies from '../components/Movies';
 
+
 function HomePage () {
-    // Returns the Movie component that displays all movies from the trending list
+    const [movies, setMovies] = useState([])
+
+    const fetchMovies = async () => {
+        const url = 'https://api.themoviedb.org/3/trending/movie/day?language=en-US';
+        const options = {
+        method: 'GET',
+        headers: {
+            accept: 'application/json',
+            Authorization: 'Bearer eyJhbGciOiJIUzI1NiJ9.eyJhdWQiOiI2MzY0NzQ0ZmM0YmM1MTkxMmFhYTQ2ZmVjZDk1OGQwNiIsInN1YiI6IjY1ZjUxNDA1Yjk3NDQyMDE2NGY5NGZmMSIsInNjb3BlcyI6WyJhcGlfcmVhZCJdLCJ2ZXJzaW9uIjoxfQ.rZbOx0atd3X7nEtWaPws8f3YenyQID1G9Hr7bbym2ek'
+        }
+        };
+
+        fetch(url, options)
+        .then(res => res.json())
+        .then(json => setMovies(json.results))
+        .catch(err => console.error('error:' + err));
+
+        console.log(movies);
+    };
+
+    useEffect(() => {
+        fetchMovies();
+      }, []);
+
     return(
         <>
-        <Movies></Movies>
+        <h1>Hello World</h1>
+        <div className='movies'>
+            {/* <div className='movie-list'>
+                {movies.map((movie) => (
+                    <>
+                    <div className='movie-details'>
+                        <Link to={`/MoviePage/${movie.id}`}>
+                            <img className='movie-img' src={`https://image.tmdb.org/t/p/w500${movie.poster_path}`} />
+                        </Link>
+                    </div>
+                    </>
+                ))}
+            </div> */}
+            <div className='movie-list'>
+                {movies.map((movie) => (
+                    <>
+                    <div className='movie-details'>
+                        <Link to={`/MoviePage/${movie.id}`}>
+                            <Movies poster={movie.poster_path} title={movie.original_title}/>
+                        </Link>
+                    </div>
+                    </>
+                ))}
+            </div>
+        </div>
         </>
     )
 }
